% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_imp.R
\name{calc_imp}
\alias{calc_imp}
\title{Calculate importance scores on a \code{st2e} object.}
\usage{
calc_imp(
  st2e_obj,
  st2e_obj2 = NULL,
  prune = T,
  prefilt_scores = NULL,
  scale = F,
  pooled = T,
  B = 1000
)
}
\arguments{
\item{st2e_obj}{A \code{st2e} object returned by \code{\link[StableMate]{st2e}}.}

\item{st2e_obj2}{A \code{st2e} object whose ensemble is the predictor pool of \code{st2e_obj}. If provided, the
importance score calculation will be adjusted accordingly.}

\item{prune}{Logical; If TRUE, weight selections by ranking quantiles of their objective scores when calculating importance scores.}

\item{prefilt_scores}{A numerical vector that contains some weighting measurements of predictor pools. Predictor pools will be
weighted by ranking quantiles of \code{prefilt_scores} when calculating importance scores.
When \code{st2e_obj2} is given, \code{prefilt_scores} will be set as the objective scores of \code{st2e_obj2}. By default is set to
a vector of ones.}

\item{scale}{Logical; If TRUE, scale the importance score of the pseudo-predictor by importance scores calculated on
predictor pools. Used only when the predictor pools of \code{st2e_obj} are pre-filtered by methods other than  \code{\link[StableMate]{st2e}}.}

\item{pooled}{Logical; If FALSE, each predictor will be benchmarked against the pseudo-predictor within the selections made
from the query predictor's pool. Otherwise, all predictors are benchmarked against the same set of pseudo-predictor selections,
which are usually all the selections in the ensemble if the pseudo-predictor is included in every predictor pool.}

\item{B}{Numeircal; The number of bootstrapping iteration for calculating importance scores.}
}
\value{
A \code{st2e} object with updated importance scores (stored in the \code{imp_scores} of the \code{st2e} object).
}
\description{
\code{\link[StableMate]{st2e}} returns a \code{st2e} object, which contains a matrix representing variable selection
ensemble and a numerical vector of objective scores corresponding to each row of the ensemble that represents one \code{st2} selection.
Importance scores are calculated as selection frequencies weighted by objective scores.
}
\seealso{
\code{\link[StableMate]{st2e}()}
}
