% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{obj_fun}
\alias{obj_fun}
\alias{obj_bic}
\alias{obj_pss}
\alias{obj_bic_logit}
\alias{obj_pss_logit}
\title{Built-in \code{st2e} objective functions}
\usage{
obj_bic(Y, X, env, pred_in)

obj_pss(Y, X, env, pred_in)

obj_bic_logit(Y, X, env, pred_in)

obj_pss_logit(Y, X, env, pred_in)
}
\arguments{
\item{Y}{A response vector or matrix depending on the objective function. It should be a vector if the function is used with
StableMate's default objective.}

\item{X}{A predictor matrix with rows representing samples and columns representing predictors. The columns must be named.}

\item{env}{A character vector indicates sample environments. Should be of the same length as the number of rows of \code{X}.}

\item{pred_in}{A indicator vector of the same length as the number of columns of \code{X}. Ones in the vector indicates the
corresponding predictors matched in column positions to be evaluated.}
}
\value{
A objective score of a minimization problem.
}
\description{
Evaluate the objective score of a given set of predictors (proposals). \code{\link[StableMate]{st2e}} algorithm
searches for the best of predictors defined by \code{\link[StableMate]{sel_fun}}. The customized objective function should
have same input as described below, and output a objective score
}
\seealso{
\code{\link[StableMate]{st2e}()}
}
