% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st2e_util.R
\name{plot.st2e}
\alias{plot.st2e}
\title{plot function for \code{st2e} object.}
\usage{
\method{plot}{st2e}(
  st2e_obj,
  imp_type = c("conditional", "joint"),
  fill_by = NULL,
  colors = NULL,
  sigthresh = 1 - exp(-5),
  plot_density = FALSE,
  bt_prop = 0.1,
  base_size = 15,
  show_labels = TRUE,
  labels = NULL,
  label_size = 10,
  box_size = 0.1
)
}
\arguments{
\item{st2e_obj}{A \code{st2e} object returned by \code{\link[StableMate]{st2e}}.}

\item{imp_type}{Either 'conditional' or 'joint' to show conditional importance scores
or joint importance scores.}

\item{fill_by}{User provided filling, should be a data frame of the same row length as the column size of the ensemble (exclude the pseudo-predictor),
Rows should be labelled by predictor names. The column name will be used as the legend title.}

\item{colors}{A vector of color values passed to \code{\link[ggplot]{scale_fill_manual}} and \code{\link[ggplot]{scale_color_manual}}.}

\item{sigthresh}{Numerical; Significance threshold.}

\item{plot_density}{Logical; If TRUE, bootstrap density of importance scores will be shown.}

\item{bt_prop}{Numerical; Proportion of bootstrap importance scores to show. Should be a value between 0 and 1.}

\item{base_size}{base_size passed to \code{\link[ggplot]{theme_classic}} .}

\item{show_labels}{Logical; If TRUE, show labels of predictors on x-axis.}

\item{labels}{User provided x label, It should be a vector named by the predictor names.}

\item{label_size}{label_size will be passed to \code{\link[ggplot]{element_text}} to control x axis label size.}

\item{box_size}{\code{\link[ggplot]{geom_boxplot}} size aesthetic.}
}
\description{
plot function for \code{st2e} object.
}
