% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemate_util.R
\name{plot.stablemate}
\alias{plot.stablemate}
\title{plot function for \code{stablemate} object.}
\usage{
\method{plot}{stablemate}(
  stbm_obj,
  pred_imp_type = c("joint", "conditional"),
  stab_imp_type = c("conditional", "joint"),
  sigthresh = 1 - exp(-5),
  color_by = NULL,
  colors = NULL,
  base_size = 15,
  point_size = 2,
  show_labels = TRUE,
  labels = NULL,
  label_size = 3,
  parse = FALSE,
  pred_cutoff = NULL,
  show_density_of = NULL
)
}
\arguments{
\item{stbm_obj}{A \code{stablemate} object returned by \code{\link[StableMate]{stablemate}}.}

\item{pred_imp_type}{Either 'conditional' or 'joint' to show conditional importance scores
or joint importance scores on predictive feature selections.}

\item{stab_imp_type}{Either 'conditional' or 'joint' to show conditional importance scores
or joint importance scores on stable feature selections.}

\item{sigthresh}{Numerical; Significance threshold.}

\item{color_by}{User provided coloring, should be a data frame of the same row length as the column size of the ensemble (exclude the pseudo-predictor),
Rows should be labelled by predictor names. The column name will be used as the legend title.}

\item{colors}{A vector of color values passed to \code{\link[ggplot]{scale_color_manual}}.}

\item{base_size}{base_size passed to \code{\link[ggplot]{theme_classic}} .}

\item{point_size}{\code{\link[ggplot]{geom_point}} size aesthetic.}

\item{show_labels}{Logical; If TRUE, show labels of predictors by \code{\link[ggrepel]{geom_text_repel}}.}

\item{labels}{User provided x label, It should be a vector named by the predictor names.}

\item{label_size}{\code{\link[ggrepel]{geom_text_repel}} size aesthetic.}

\item{parse}{Logical; If TRUE, phrase rich-text expressions passed to \code{labels}.}

\item{show_density_of}{A character vector of names of predictors to show bootstrap importance density of.}

\item{plot_density}{Logical; If TRUE, bootstrap density of importance scores will be shown.}

\item{bt_prop}{Numerical; Proportion of bootstrap importance scores to show. Should be a value between 0 and 1.}

\item{pred_cutoff;}{A numerical between 0 and 1. Predictors with predictivity scores larger than
\code{pred_cutoff} will be colored and labelled.}
}
\description{
plot function for \code{stablemate} object.
}
