% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st2e_util.R
\name{predict.st2e}
\alias{predict.st2e}
\title{predict function for \code{st2e} object.}
\usage{
\method{predict}{st2e}(st2e_obj, X, prune = TRUE, prefilt_scores = NULL, pred_fun = pred_ols)
}
\arguments{
\item{st2e_obj}{A \code{st2e} object returned by \code{\link[StableMate]{st2e}}.}

\item{X}{New predictor data that can be recognized by \code{st2e} regression models fitted by the \code{reg_fun} argument passed to \code{\link[StableMate]{st2e}}.}

\item{prune}{Logical; If TRUE, weight selections by ranking quantiles of their objective scores when aggregating regression models.}

\item{prefilt_scores}{A numerical vector that contains some weighting measurements of predictor pools. Predictor pools will be
weighted by ranking quantiles of \code{prefilt_scores} when aggregating regression models.}

\item{pred_fun}{The function that makes prediction on the response given new predictor data \code{X} and regression models
in \code{st2e}. By default, expect regression models fitted by code{\link[StableMate]{reg_ols}}.
See \code{\link[StableMate]{reg_fun}} for more options and how to customize this function.}
}
\description{
predict function for \code{st2e} object.
}
