% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemate_util.R
\name{print.stablemate}
\alias{print.stablemate}
\title{print function for \code{stablemate} object.}
\usage{
\method{print}{stablemate}(
  stbm_obj,
  pred_imp_type = c("joint", "conditional"),
  stab_imp_type = c("conditional", "joint"),
  sigthresh = 1 - exp(-5)
)
}
\arguments{
\item{stbm_obj}{A \code{stablemate} object returned by \code{\link[StableMate]{stablemate}}.}

\item{pred_imp_type}{Either 'conditional' or 'joint' to make predictive feature selections
based on conditional importance scores or joint importance scores.}

\item{stab_imp_type}{Either 'conditional' or 'joint' to make stable feature selections
based on conditional importance scores or joint importance scores.}

\item{sigthresh}{Numerical; Significance threshold.}
}
\description{
print function for \code{stablemate} object.
}
