% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{reg_fun}
\alias{reg_fun}
\alias{reg_ols}
\alias{reg_logit}
\title{Built-in \code{st2e} regression functions.}
\usage{
reg_ols(Y, X, env, pred_in)

reg_logit(Y, X, env, pred_in)
}
\arguments{
\item{Y}{A response vector or matrix depending on the objective function. It should be a vector if the function is used with
StableMate's default objective.}

\item{X}{A predictor matrix with rows representing samples and columns representing predictors. The columns must be named.}

\item{env}{A character vector indicates sample environments. Should be of the same length as the number of rows of \code{X}.}

\item{pred_in}{A indicator vector of the same length as the number of columns of \code{X}. Ones in the vector indicates the
corresponding predictors matched in column positions to be fitted.}
}
\value{
A regression model.
}
\description{
Build regression models based on the optimal set of predictors (optimal proposal) after \code{\link[StableMate]{st2e}} search.
The customized regression function should have same input as described below, and output a regression model that can be
used to make predictions.
}
\seealso{
\code{\link[StableMate]{st2e}()}
}
