% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{sel_fun}
\alias{sel_fun}
\alias{sel_min}
\title{Built-in \code{st2e} proposal selection functions.}
\usage{
sel_min(scores)
}
\arguments{
\item{scores}{A numerical vector of objective scores.}
}
\value{
The index of the best proposal.
}
\description{
Given objective scores of multiple proposals calculated by objective functions
 \code{\link[StableMate]{obj_fun}} provided to \code{\link[StableMate]{st2e}},
defines how to select the best proposal to make a step. The customized selection function should have same input as described below,
and output the index of the best proposal.
}
\seealso{
\code{\link[StableMate]{obj_fun}}
}
