% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{sim_scm}
\alias{sim_scm}
\title{Simulate a directed asymmetric graph.}
\usage{
sim_scm(nSample = 300, nE = 4, nNode = 50, propE = 0.1, nI = 30, idxY = 25)
}
\arguments{
\item{nSample}{Numerical; Number of samples to simulate per environments.}

\item{nE}{Numerical; Number of environments to simulate.}

\item{nNode}{Numerical; Number of nodes (one response + predictors) in the system.}

\item{propE}{Numerical; Proportion of edges to be connected.}

\item{nI}{Numerical; A number between 1 and \code{nNode} controls the number of predictors to be intervened.}

\item{idXY}{Numerical; A number between 1 and \code{nNode} represents a variable index to be chosen as the response}
}
\value{
A \code{list} contains simulation results.
}
\description{
Simulate a directed asymmetric graph that satisfies the model assumption of
stabilized regression (\href{https://doi.org/10.1214/21-AOAS1487}{Pfister et al. (2019, The Annals of Applied Statistics)}).
}
