% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st2e.R
\name{st2e}
\alias{st2e}
\alias{st2_fwd}
\alias{st2_bwd}
\alias{st2_swt}
\title{Stochastic stepwise variable selection ensemble}
\usage{
st2e(
  Y,
  X,
  env = NULL,
  obj_fun = obj_bic,
  sel_fun = sel_min,
  reg_fun = reg_ols,
  pred_pool = NULL,
  start_set = "random_start",
  max_size = NULL,
  alpha = 1,
  beta = 5,
  lambda = 0.5,
  a = 1,
  b = 1,
  t = 0,
  K = 100,
  do_switch = FALSE,
  ret_mod = TRUE,
  ret_imp = TRUE,
  calc_imp_ctrl = NULL,
  ncore = NULL,
  par_method = c("SNOW", "MC", "MPI"),
  verbose = TRUE,
  chunk_size = ceiling(K/ncore),
  fun_export = NULL,
  pkg_export = NULL,
  drop_pred = TRUE
)

st2_fwd(
  Y,
  X,
  env,
  obj_fun,
  sel_fun,
  pred_in,
  pred_pool,
  max_size,
  alpha,
  beta,
  lambda,
  a,
  b,
  t,
  fixed_size = F
)

st2_bwd(
  Y,
  X,
  env,
  obj_fun,
  sel_fun,
  pred_in,
  pred_pool,
  max_size,
  alpha,
  beta,
  lambda,
  a,
  b,
  t
)

st2_swt(
  Y,
  X,
  env,
  obj_fun,
  sel_fun,
  pred_in,
  pred_pool,
  max_size,
  alpha,
  beta,
  lambda,
  a,
  b,
  t
)
}
\arguments{
\item{Y}{A response vector or matrix depending on the objective function. It should be a vector if the function is used with
StableMate's default objective.}

\item{X}{A predictor matrix with rows representing samples and columns representing predictors. The columns must be named.}

\item{obj_fun}{The function that calculates objective scores for any given set of predictors. See our default objective function to learn the syntax. By
default the objective is to minimize linear regression BIC scores. See \code{\link[StableMate]{obj_fun}} for the other options and how to customize this function.}

\item{sel_fun}{The function that selects the best set of st2e proposal given multiple proposals and their objective scores. By
default we select the proposal with the minimum objective score. See \code{\link[StableMate]{sel_fun}} for how to customize this function.}

\item{reg_fun}{The regression function. By default the we perform ordinary least square regression.
See \code{\link[StableMate]{reg_fun}} for the other options and how to customize this function.}

\item{pred_pool}{A vector or a matrix that restricts the variable selection space. If a vector is given, it must be a vector of characters that is a subset of
the column names of X. Section of "pseudo_predictor" can be enabled by adding a "pseudo_predictor" into \code{pred_pool}. When a matrix is given, it must be an indicator matrix
of \code{K} rows with column names being a subset of the column names of x. Section of "pseudo_predictor" can be enabled by adding a "pseudo_predictor" column into \code{pred_pool}.}

\item{start_set}{A vector or a matrix that controls the starting point of st2e selection. If a vector is given, it must be a vector of characters that is a subset of \code{pred_pool}
When a matrix is given, it must be an indicator matrix of \code{K} rows with column names being a subset of the column names of \code{pred_pool}. \code{start_set} must be smaller or
equal to \code{pred_pool} in every elements of their matching column. By default, we start the selection from a random predictor set under \code{pred_pool}.}

\item{max_size}{Numerical; Any step of selection will not consider a predictor set that is larger than \code{max_size}.}

\item{alpha}{Numerical; A Parameter of beta distribution for sampling proposal size.}

\item{beta}{Numerical; A Parameter of beta distribution for sampling proposal size.}

\item{lambda}{Numerical; A scaling factor that controls the size of binomial distribution for sampling proposal size.. Should be a value between 0 and 1.
If 1, then the proposal size can be as large as the size of the maximum possible proposal.}

\item{a}{Numerical; A scaling factor that controls the number of proposal to sample, which contains a term where \code{a} is multiplied by the size of the maximum possible proposal
to the power of \code{t}.}

\item{b}{Numerical; A scaling factor that controls the number of proposal to sample, which contains a term where \code{b} is multiplied by a quality that measures the number of
all possible unique proposal that can be made.}

\item{t}{Numerical; A scaling factor that controls the number of proposal to sample, which contains a term where the size of the maximum possible proposal is rasied to the power of
\code{t} and multiplied by \code{a}.}

\item{K}{Numerical; Ensemble size.}

\item{do_switch}{Logical; If true, switch steps that swap predictors that are in and out the model are allowed.}

\item{ret_mod}{Logical; If true, return the prediction ensemble}

\item{ret_imp}{Logical; If true, return the importance score.}

\item{calc_imp_ctrl}{A list that controls importance score calculation. See the document of \code{calc_imp} for details.}

\item{ncore}{Numerical; Numerical; If greater than 0. Parallel computing is enabled.}

\item{verbose}{Logical; If true, multi-core verbose will be printed.}

\item{chunk_size}{Numerical; The size of task chunks (How many st2 runs are performed one each task).}

\item{fun_export}{A vector of names of functions to be imported by \code{foreach} environment.}

\item{pkg_export}{A vector of names of packages to be imported by \code{foreach} environment.}

\item{drop_pred}{Logical; If true, remove predictors from X that are not in the predictor pool to save space and time.}

\item{par_method;}{Parallel computing method. SNOW is preferred on Windows local machines, MC is preferred on non-Windows local machines.
MPI is preferred on distributed computing cloud.}
}
\value{
A \code{st2e} object.
}
\description{
The main st2e function for stochastic stepwise variable selection proposed by
\href{https://www.tandfonline.com/doi/full/10.1080/10618600.2012.679223}{Lu Xin &Mu Zhu (2010, Journal of Computational and Graphical Statistics)}.
}
