% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemate.R
\name{stablemate}
\alias{stablemate}
\title{StableMate regression and stability analysis}
\usage{
stablemate(
  Y,
  X,
  env,
  K = 100,
  alpha = 1,
  beta = 5,
  lambda = 0.5,
  a = 1,
  b = 1,
  t = 0,
  max_size = NULL,
  pred_st2_ctrl = NULL,
  stab_st2_ctrl = NULL,
  pred_calc_imp_ctrl = NULL,
  stab_calc_imp_ctrl = NULL,
  do_switch = FALSE,
  ret_mod = TRUE,
  ret_imp = TRUE,
  ncore = NULL,
  par_method = c("SNOW", "MC", "MPI"),
  verbose = TRUE,
  chunk_size = ceiling(K/ncore),
  fun_export = NULL,
  pkg_export = NULL,
  drop_pred = TRUE
)
}
\arguments{
\item{Y}{A response vector or matrix depending on the objective function. It should be a vector if the function is used with
StableMate's default objective.}

\item{X}{A predictor matrix with rows representing samples and columns representing predictors. The columns must be named.}

\item{env}{A character vector indicates sample environments. Should be of the same length as the number of rows of \code{X}.}

\item{K}{Numerical; Ensemble size.}

\item{alpha}{Numerical; A Parameter of beta distribution for sampling proposal size.}

\item{beta}{Numerical; A Parameter of beta distribution for sampling proposal size.}

\item{lambda}{Numerical; A scaling factor that controls the size of binomial distribution for sampling proposal size.. Should be a value between 0 and 1.
If 1, then the proposal size can be as large as the size of the maximum possible proposal.}

\item{a}{Numerical; A scaling factor that controls the number of proposal to sample, which contains a term where \code{a} is multiplied by the size of the maximum possible proposal
to the power of \code{t}.}

\item{b}{Numerical; A scaling factor that controls the number of proposal to sample, which contains a term where \code{b} is multiplied by a quality that measures the number of
all possible unique proposal that can be made.}

\item{t}{Numerical; A scaling factor that controls the number of proposal to sample, which contains a term where the size of the maximum possible proposal is raised to the power of
\code{t} and multiplied by \code{a}.}

\item{max_size}{Numerical; Any step of selection will not consider a predictor set that is larger than \code{max_size}.}

\item{pred_st2_ctrl}{Controls for \code{\link[StableMate]{st2e}} that selects most predictive features of the response. See \code{\link[StableMate]{st2e}} for the detailed description of controls.
\code{K}, \code{ret_imp}, \code{calc_imp_ctrl}, \code{drop_pred} will be ignored if provided.}

\item{stab_st2_ctrl}{Controls for \code{\link[StableMate]{st2e}} that selects most stable preidctors of the response. See \code{\link[StableMate]{st2e}} for the detailed description of controls.
\code{K}, \code{ret_imp}, \code{calc_imp_ctrl}, \code{drop_pred} will be ignored if provided. By default, stable predictors will be searched within predictive features.}

\item{pred_calc_imp_ctrl}{Controls for \code{\link[StableMate]{calc_imp}} applied on the predictive ensemble for calculating predictivity scores.
See \code{\link[StableMate]{calc_imp}} for the detailed description of controls.}

\item{stab_calc_imp_ctrl}{Controls for \code{\link[StableMate]{calc_imp}} applied on the stable and predictive ensemble for calculating stability scores.
See \code{\link[StableMate]{calc_imp}} for the detailed description of controls.  By defaults, selections in the stable and predictive ensemble
are considered as the subsets of selections in the predictive ensemble, and stability scores are adjusted accordingly.}

\item{do_switch}{Logical; If true, switch steps that swap predictors that are in and out the model are allowed.}

\item{ret_mod}{Logical; If true, return the prediction ensemble}

\item{ret_imp}{Logical; If true, return the importance score.}

\item{ncore}{Numerical; Numerical; If greater than 0. Parallel computing is enabled.}

\item{verbose}{Logical; If true, multi-core verbose will be printed.}

\item{chunk_size}{Numerical; The size of task chunks (How many st2 runs are performed one each task).}

\item{fun_export}{A vector of names of functions to be imported by \code{\link[foreach]{foreach}} environment.}

\item{pkg_export}{A vector of names of packages to be imported by\code{\link[foreach]{foreach}} environment.}

\item{drop_pred}{Logical; If true, remove predictors from X that are not in the predictor pool to save space and time.}

\item{par_method;}{Parallel computing method. SNOW is preferred on Windows local machines, MC is preferred on non-Windows local machines.
MPI is preferred on distributed computing cloud.}
}
\value{
A \code{stablemate} object.
}
\description{
Perform stabilized regression (\href{https://doi.org/10.1214/21-AOAS1487}{Pfister et al. (2019, The Annals of Applied Statistics)})
by stochastic stepwise variable selection. Many controls in this function are shared by the \code{\link[StableMate]{st2e}}.
}
